requirements.txt
streamlit
pandas
requests
websocket-client
matplotlib
Pillow
